020_dbd_mysql.dpatch, 020_dbd_mysql_autofoo.dpatch: Add external module apr_dbd_mysql.c, now that it has been relicensed to be compatible with the Apache License 2.0.  Thanks to Nick Kew for working with us on this.  This is for bugs #395959 / #405773, allowing apache2 to use mysql for auth.
The main consequence is that users of apr-util, notably apache2 and
subversion, now also pull in libmysqlclient15off, which is 4 MB or so.
It may be possible to split the dbd backends into loadable modules, so
that such libraries can be Recommends instead, but I don't feel
comfortable doing that for etch.


git-svn-id: svn+ssh://svn.debian.org/svn/pkg-apache/trunk/apr-util@295 01b336ce-410b-0410-9a02-a0e7f243c266